Add  Video - Audio bitrate per channel
Adds option to change bitrate of audio streams based on a number of channel basis to preserve current mix where possible. Works similarly to the -vbr setting of libfdk_aac and similar options of other codecs. Includes fallback option for when channels are not available in stream's metadata. Also includes optional buffer to keep existing audio/settings if it is within a range of the original track